Misc Cleanup
* Removes the use of ALEMBIC_MAYA_PLUGIN_EXPORT and Export.h
* Adds the override keyword to overriden methods
* Cleans up some compiler warnings

+++ maya/AbcExport/MayaNurbsCurveWriter.cpp
@@ -296,27 +296,27 @@ void MayaNurbsCurveWriter::write()
             if (knotsArray[0] == knotsArray[knotsLength - 1] ||
                 knotsArray[0] == knotsArray[1])
             {
-                knots.push_back(knotsArray[0]);
+                knots.push_back(static_cast<float>(knotsArray[0]));
             }
             else
             {
-                knots.push_back(2 * knotsArray[0] - knotsArray[1]);
+                knots.push_back(static_cast<float>(2 * knotsArray[0] - knotsArray[1]));
             }
 
             for (unsigned int j = 0; j < knotsLength; ++j)
             {
-                knots.push_back(knotsArray[j]);
+                knots.push_back(static_cast<float>(knotsArray[j]));
             }
 
             if (knotsArray[0] == knotsArray[knotsLength - 1] ||
                 knotsArray[knotsLength - 1] == knotsArray[knotsLength - 2])
             {
-                knots.push_back(knotsArray[knotsLength - 1]);
+                knots.push_back(static_cast<float>((knotsArray[knotsLength - 1])));
             }
             else
             {
-                knots.push_back(2 * knotsArray[knotsLength - 1] -
-                                knotsArray[knotsLength - 2]);
+                knots.push_back(static_cast<float>(2 * knotsArray[knotsLength - 1] -
+                                knotsArray[knotsLength - 2]));
             }
         }
